+++ src/utils.c
@@ -100,10 +100,10 @@ void utils_clear_buffers(void)
  *
  * @return Nothing.
  */
-void utils_hex_to_bin(const char* str, unsigned char* out, int inLen, int* outLen)
+void utils_hex_to_bin(const char* str, unsigned char* out, size_t inLen, size_t* outLen)
     {
-    int bLen = inLen / 2;
-    int i;
+    size_t bLen = inLen / 2;
+    size_t i;
     dogecoin_mem_zero(out, bLen);
     for (i = 0; i < bLen; i++) {
         if (str[i * 2] >= '0' && str[i * 2] <= '9') {
@@ -230,10 +230,10 @@ char* utils_uint8_to_hex(const uint8_t* bin, size_t l)
  *
  * @return Nothing.
  */
-void utils_reverse_hex(char* h, int len)
+void utils_reverse_hex(char* h, size_t len)
     {
     char* copy = dogecoin_calloc(1, len);
-    int i;
+    size_t i;
     memcpy_safe(copy, h, len);
     for (i = 0; i < len - 1; i += 2) {
         h[i] = copy[len - i - 2];
@@ -351,10 +351,10 @@ void* safe_malloc(size_t size)
  *
  * @return Nothing.
  */
-void dogecoin_cheap_random_bytes(uint8_t* buf, uint32_t len)
+    void dogecoin_cheap_random_bytes(uint8_t* buf, size_t len)
     {
     srand(time(NULL)); // insecure
-    for (uint32_t i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
         buf[i] = rand(); // weak non secure cryptographic rng
         }
     }
@@ -484,7 +484,7 @@ void prepend(char* s, const char* t)
  */
 void append(char* s, char* t)
     {
-    int i = 0, length = 0;
+    size_t i = 0, length = 0;
     /* get length of char* s */
     for (; memcmp(&s[i], "\0", 1) != 0; i++) length++;
 
@@ -503,8 +503,8 @@ void append(char* s, char* t)
  * @param output
  */
 void text_to_hex(char* in, char* out) {
-    int length = 0;
-    int i = 0;
+    size_t length = 0;
+    size_t i = 0;
 
     while (in[length] != '\0') {
         sprintf((char*)(out + i), "%02X", in[length]);
@@ -531,7 +531,7 @@ const char* get_build() {
 /* reverse:  reverse string s in place */
 void dogecoin_str_reverse(char s[])
 {
-    int i, j;
+    size_t i, j;
     char c;
 
     for (i = 0, j = strlen(s)-1; i<j; i++, j--) {
