+++ test/opreturn_tests.c
@@ -38,7 +38,7 @@ void test_op_return() {
     // convert message text to hexadecimal:
     text_to_hex(msg, (char*)msg_hex);
 
-    int length = (strlen((char*)msg_hex) / 2) + 1; // 69
+    size_t length = (strlen((char*)msg_hex) / 2) + 1; // 69
     
     char decimal[32];
     sprintf(decimal, "%x", length - 1);
@@ -46,7 +46,7 @@ void test_op_return() {
 
     u_assert_int_eq(memcmp(msg_hex, expected_hexmsg, strlen(expected_hexmsg)), 0);
 
-    int outlen;
+    size_t outlen;
     uint8_t* script_data=dogecoin_uint8_vla(strlen((char*)msg_hex));
     utils_hex_to_bin((char*)msg_hex, script_data, strlen((char*)msg_hex), &outlen);
